+++ packages/@expo/cli/src/start/server/metro/debugging/messageHandlers/VscodeDebuggerSetBreakpointByUrl.ts
@@ -1,7 +1,8 @@
-import Protocol from 'devtools-protocol';
+import type Protocol from 'devtools-protocol';
 
-import { CdpMessage, DebuggerMetadata, DebuggerRequest, InspectorHandler } from './types';
-import { getDebuggerType } from './utils';
+import { MessageHandler } from '../MessageHandler';
+import { getDebuggerType } from '../getDebuggerType';
+import type { CdpMessage, DebuggerRequest } from '../types';
 
 /**
  * Hermes and vscode have trouble setting breakpoints by `urlRegex` through `Debugger.setBreakpointByUrl`.
@@ -9,16 +10,13 @@ import { getDebuggerType } from './utils';
  * Hermes needs to create the breakpoint to get the proper ID, but it must be unbounded.
  * Once the sourcemap is loaded, vscode will rebind the unbounded breakpoint to the correct location (using `Debugger.setBreakpoint`).
  */
-export class VscodeDebuggerSetBreakpointByUrlHandler implements InspectorHandler {
-  onDebuggerMessage(
-    message: DebuggerRequest<DebuggerSetBreakpointByUrl>,
-    { userAgent }: DebuggerMetadata
-  ): boolean {
-    if (
-      getDebuggerType(userAgent) === 'vscode' &&
-      message.method === 'Debugger.setBreakpointByUrl' &&
-      message.params.urlRegex
-    ) {
+export class VscodeDebuggerSetBreakpointByUrlHandler extends MessageHandler {
+  isEnabled() {
+    return getDebuggerType(this.debugger.userAgent) === 'vscode';
+  }
+
+  handleDebuggerMessage(message: DebuggerRequest<DebuggerSetBreakpointByUrl>) {
+    if (message.method === 'Debugger.setBreakpointByUrl' && message.params.urlRegex) {
       // Explicitly force the breakpoint to be unbounded
       message.params.url = 'file://__invalid_url__';
       delete message.params.urlRegex;